example_streamlit/requirements.txt
streamlit==1.39.0
pandas==2.2.3
psycopg2-binary==2.9.10
folium==0.18.0
streamlit-folium==0.23.1
